Quick conversion fix

+++ FIRMWARE/TER_PEDAL/Core/Inc/TeR_UTILS.h
@@ -11,7 +11,7 @@
 #include "stm32f1xx_hal.h"
 
 //Voltage to ADC value
-#define VOLT2ADC(V) ((uint32_t)(((V) / 3.3) * 4095))
+#define VOLT2ADC(V) ((((V) / 3.3) * 4095.0))
